\name{data1}
\alias{data1}
\docType{data}
\title{
data1
}
\description{
Dataset used to demonstrate the use of ab-SNF. 
}
\usage{
data(data1)
}
\format{
A data frame with 150 observations on the following 1000 variables.\cr
Refer to senario 1(2) in simulation part of WUsing Association Signal Annotations to boost Similarity Network Fusion (2018), Peifeng Ruan, Ya Wang, Ronglai Shen, Shuang Wang.

}
\examples{
data(data1)
}
\keyword{datasets}
