\name{data2}
\alias{data2}
\docType{data}
\title{
data2
}
\description{
Dataset used to demonstrate the use of ab-SNF. 
}
\usage{
data(data2)
}
\format{
A data frame with 150 observations on the following 1000 variables.\cr
Refer to continuous features in senario 1(2) in simulation part of Using Association Signal Annotations to boost Similarity Network Fusion(2018), Peifeng Ruan, Ya Wang, Ronglai Shen, Shuang Wang.

  }

\examples{
data(data2)
}
\keyword{datasets}
