\name{abSNF-internal} 
\alias{.csPrediction} 
\alias{.discretisation} 
\alias{.discretisationEigenVectorData} 
\alias{.dominateset} 
\alias{.mutualInformation} 
\alias{.entropy} 
\title{Internal abSNF Functions} 
\description{ 
          Internal abSNF functions 
} 
\details{ 
          These are not to be called by the user. 
} 
\keyword{ internal } 
