\name{true_label}
\alias{true_label}
\docType{data}
\title{
true_label
}
\description{
Dataset used to demonstrate the use of abSNF. 
}
\usage{
data(true_label)
}
\format{
True group lables for 150 subjects.\cr
Refer to senario 1(2) in simulation part of Using Association Signal Annotations to boost Similarity Network Fusion(2018), Peifeng Ruan, Ya Wang, Ronglai Shen, Shuang Wang.

  }

\examples{
data(true_label)
}
\keyword{datasets}
