\name{weight1}
\alias{weight1}
\docType{data}
\title{
weight1
}
\description{
Dataset used to demonstrate the use of abSNF. 
}
\usage{
data(weight1)
}
\format{
A sequence indicating the weights for 1,000 features in data1.\cr
Refer to continuous features in senario 1(2) in simulation part of Using Association Signal Annotations to boost Similarity Network Fusion(2018), Peifeng Ruan, Ya Wang, Ronglai Shen, Shuang Wang.

  }

\examples{
data(weight1)
}
\keyword{datasets}
