\name{weight2}
\alias{weight2}
\docType{data}
\title{
weight2
}
\description{
Dataset used to demonstrate the use of abSNF. 
}
\usage{
data(weight2)
}
\format{
A sequence indicating the weights for 1,000 features in data2.\cr
Refer to senario 1(2) in simulation part of Using Association Signal Annotations to boost Similarity Network Fusion(2018), Peifeng Ruan, Ya Wang, Ronglai Shen, Shuang Wang.

  }

\examples{
data(weight2)
}
\keyword{datasets}
